Move multiblock/master/server logic to separate package
