+++ src/main/java/slimeknights/mantle/multiblock/IServantLogic.java
@@ -1,4 +1,4 @@
-package slimeknights.mantle.block;
+package slimeknights.mantle.multiblock;
 
 import net.minecraft.util.BlockPos;
 import net.minecraft.world.World;
